add OPENVINO_GENAI_C_EXPORTS and regulate folder layer
