+++ src/cpp/include/openvino/genai/visibility.hpp
@@ -7,6 +7,8 @@
 
 #ifdef openvino_genai_EXPORTS
 #    define OPENVINO_GENAI_EXPORTS OPENVINO_CORE_EXPORTS
+#    define OPENVINO_GENAI_C_EXPORTS OPENVINO_EXTERN_C OPENVINO_CORE_EXPORTS
 #else
 #    define OPENVINO_GENAI_EXPORTS OPENVINO_CORE_IMPORTS
+#    define OPENVINO_GENAI_C_EXPORTS OPENVINO_EXTERN_C OPENVINO_CORE_IMPORTS
 #endif  // openvino_genai_EXPORTS
